clawAgent/pyproject.toml
[project]
name = "clawagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.28.0",
    "openrouter>=0.7.11",
    "python-dotenv>=1.2.2",
]
